+++ runtime/src/pallets/mod.rs
@@ -1,0 +1,56 @@
+// This file is part of Darwinia.
+//
+// Copyright (C) 2018-2022 Darwinia Network
+// SPDX-License-Identifier: GPL-3.0
+//
+// Darwinia is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Darwinia is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Darwinia. If not, see <https://www.gnu.org/licenses/>.
+
+mod system;
+pub use system::*;
+
+mod timestamp;
+pub use timestamp::*;
+
+mod authorship;
+pub use authorship::*;
+
+mod balances;
+pub use balances::*;
+
+mod transaction_payment;
+pub use transaction_payment::*;
+
+mod parachain_system;
+pub use parachain_system::*;
+
+mod parachain_info_;
+pub use parachain_info_::*;
+
+mod aura_ext;
+pub use aura_ext::*;
+
+mod xcmp_queue;
+pub use xcmp_queue::*;
+
+mod dmp_queue;
+pub use dmp_queue::*;
+
+mod session;
+pub use session::*;
+
+mod aura;
+pub use aura::*;
+
+mod collator_selection;
+pub use collator_selection::*;
